Modify the libtool transformation script to attempt to fix the OpenBSD buildbot.
* build/generator/gen_make.py
  (write_transform_libtool_scripts): Look for .so.0 and .so.0.0, don't
   write LD_PRELOAD lines if empty.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1050332 13f79535-47bb-0310-9956-ffa450edef68

+++ build/generator/gen_make.py
@@ -493,24 +493,34 @@ transform()
         case $LIB in
           *libsvn_test-*) continue ;;
         esac
-        if [ -f $LIB ]; then
-          if [ -z "$EXISTINGLIBS" ]; then
-            EXISTINGLIBS="$LIB"
-          else
-            EXISTINGLIBS="$EXISTINGLIBS $LIB"
-          fi
+        if [ ! -f $LIB ]; then
+           LIB=${LIB}.0
+           if [ ! -f $LIB ]; then
+             LIB=${LIB}.0
+             if [ ! -f $LIB ]; then
+               continue
+             fi
+           fi
+        fi
+
+        if [ -z "$EXISTINGLIBS" ]; then
+          EXISTINGLIBS="$LIB"
+        else
+          EXISTINGLIBS="$EXISTINGLIBS $LIB"
         fi
       done
-      cat "$SCRIPT" |
-      (
-        read LINE
-        echo "$LINE"
-        echo "LD_PRELOAD=\\"$EXISTINGLIBS\\""
-        echo "export LD_PRELOAD"
-        cat
-      ) < "$SCRIPT" > "$SCRIPT.new"
-      mv -f "$SCRIPT.new" "$SCRIPT"
-      chmod +x "$SCRIPT"
+      if [ ! -z "$EXISTINGLIBS" ]; then
+        cat "$SCRIPT" |
+        (
+          read LINE
+          echo "$LINE"
+          echo "LD_PRELOAD=\\"$EXISTINGLIBS\\""
+          echo "export LD_PRELOAD"
+          cat
+        ) < "$SCRIPT" > "$SCRIPT.new"
+        mv -f "$SCRIPT.new" "$SCRIPT"
+        chmod +x "$SCRIPT"
+      fi
     fi
   fi
 }
